+++ scripts/build_install_iso/03_general_preparation.sh
@@ -85,7 +85,7 @@ create_a_new_setup() {
 #   MAIN CHECKING SOME LIBRARY CONSISTENCY
 #************************************************************************************************************************
 
-if [ -z "$BASH" ]; then printf "\nWRONG SHELL: '%s' SCRIPT needs 'bash'\n\n" "$(ps -p $$ -ocomm=)"; exit 1; fi
+if [ -z "$BASH" ]; then printf "\nERROR SHELL: '%s' SCRIPT needs 'bash'\n\n" "$(ps -p $$ -ocomm=)"; exit 1; fi
 THIS_SCRIPT_PATH="$(readlink -f "$(type -P $0 || echo $0)")"
 THIS_SCRIPT_DIR="$(dirname "$THIS_SCRIPT_PATH")"
 source $THIS_SCRIPT_DIR/../p_linux_common_func.sh